ajout des autorisations sur le CRUD

+++ controllers/postController.js
@@ -3,7 +3,7 @@ const validationHandler = require('../validations/validationHandler');
 
 exports.index = async (req, res) => {
   try {
-      const posts = await Post.find().sort({ createAt: -1});
+      const posts = await Post.find().populate('user').sort({ createAt: -1});
       res.send(posts);
   } catch (err) {
     next(err);
@@ -13,7 +13,7 @@ exports.show = async (req, res, next) => {
   try {
     const post = await Post.findOne({
       _id:req.params.id
-    })
+    }).populate('user')
     res.send(post);
   } catch (err) {
     next(err);
@@ -26,6 +26,7 @@ exports.store =  async (req, res, next) => {
     let post = new Post();
     post.description = req.body.description;
     post.image = req.file.filename;
+    post.user = req.user
     post = await post.save();
 
     res.send(post);
@@ -37,7 +38,13 @@ exports.update =  async (req, res, next) => {
   try {
     validationHandler(req);
 
-    let post = await Post.findById(req.params.id); 
+    let post = await Post.findById(req.params.id);
+    // Savoir si le current user est l'auteur
+    if (!post || post.user != req.user.id) {
+      const error = new Error("Wrong request");
+      error.statusCode = 400;
+      throw error;
+    } 
     post.description = req.body.description;
     post = await post.save();
 
@@ -51,6 +58,11 @@ exports.delete =  async (req, res, next) => {
     validationHandler(req);
 
     let post = await Post.findById(req.params.id); 
+    if (!post || post.user != req.user.id) {
+      const error = new Error("Wrong request");
+      error.statusCode = 400;
+      throw error;
+    } 
     await post.delete();
 
     res.send({message: 'success'});
